feat: add a language usage check for users before deleting the language

+++ src/modules/languages/languages.service.ts
@@ -1,13 +1,17 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, forwardRef, Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { ObjectId, FilterQuery } from 'mongoose';
 import { LANGUAGE_EXCEPTION_MESSAGE } from 'src/common/constants';
+import { UsersService } from '../users/users.service';
 import { AllLanguagesResponseDto, CreateLanguageDto, GetAllLanguagesQueryDto, UpdateLanguageDto } from './dto';
 import { Language } from './language.schema';
 import { LanguagesRepository } from './languages.repository';
 
 @Injectable()
 export class LanguagesService {
-  constructor(private languagesRepository: LanguagesRepository) {}
+  constructor(
+    private languagesRepository: LanguagesRepository,
+    @Inject(forwardRef(() => UsersService)) private usersService: UsersService,
+  ) {}
 
   findAndCountAll = async (query: GetAllLanguagesQueryDto): Promise<AllLanguagesResponseDto> => {
     const languagesAndTheirCount = await this.languagesRepository.findAndCountAll(query);
@@ -46,6 +50,11 @@ export class LanguagesService {
     const languageToDelete = await this.findOne({ _id: id });
     if (!languageToDelete) {
       throw new NotFoundException(LANGUAGE_EXCEPTION_MESSAGE.NOT_FOUND);
+    }
+
+    const languageIsUsedInUsers = await this.usersService.findOne({ nativeLanguageId: id });
+    if (languageIsUsedInUsers) {
+      throw new BadRequestException(LANGUAGE_EXCEPTION_MESSAGE.CANNOT_BE_DELETED);
     }
 
     await this.languagesRepository.delete(id);
